Minor type-related improvements

+++ Kernel/Tactic.cpp
@@ -1664,7 +1664,7 @@ bool DlSatTester :: commonTacticBodyProj ( const TRole* R, BipolarPointer C, con
 	// checkProjection() might change curNode's edge vector and thusly invalidate iterators
 	DlCompletionTree::const_edge_iterator p = curNode->begin(), p_end = curNode->end();
 
-	for ( long i = 0, n = p_end - p; i < n; ++i )
+	for ( ptrdiff_t i = 0, n = p_end - p; i < n; ++i )
 	{
 		p = curNode->begin() + i;
 		if ( (*p)->isNeighbour(R) )
